Add check for USE_HSPLL definition when calculating serial baud rate
In MPLAB8, the only config word option for the oscillator which will work
for me to allow debugging a PIC18F452 with an ICD2 is the "HS with PLL"
option. That option makes the system frequency 4x the oscillator frequency.
Therefore this must be accounted for when determining the UART baud rate
generator value.
In such a setup, define USE_HSPLL in the CFLAGS when compiling.