Move 8255 control word changes to ata reg access functions
Changing the port directions in the 8255 control word causes affected ports
to be reset to output all 0s. This can have adverse effects when switching
direction on a port in the middle of an ATA read transaction, particularly
with the ATA address and control signals which are output on one of the
8255 ports being cleared whilst still in use.
So changing the control word is now performed at the start of an ATA read
or write, in the ATA register access functions, and is neatly abstracted by
a single call to set8255DataPortsDir().

Untested whether this gets it all working or not.